Fixes PRNG and implement seed interface for runner

+++ inkcpp/numeric_operations.h
@@ -402,7 +402,7 @@ namespace ink::runtime::internal {
 		void operator()(basic_eval_stack& stack, value* vals) {
 			int min = casting::numeric_cast<value_type::int32>(vals[0]);
 			int max = casting::numeric_cast<value_type::int32>(vals[0]);
-			stack.push(value{}.set<value_type::int32>(_prng.rand(max - min + 1) + min));
+			stack.push(value{}.set<value_type::int32>(static_cast<int32_t>(_prng.rand(max - min + 1) + min)));
 		}
 	};
 }
